Expand on explanation of drop fields for line charts and area charts
As a better partial fix, we want to change the logic from dropping quant fields to dropping x and y fields and showing only other fields that are used for encoding for area and line charts. In this way, if a qualitative field is NOT used for encoding, it will NOT show up in the tooltip of a line/area chart.