Check shouldShowTooltip on tooltip init, update and clear
Before this commit we only check shouldShowTooltip on init, but actually we should check should check this on init, update and clear. If a tooltip should not be init in the first place, then it should not be updated or cleared. This makes sure that user-provided callbacks don't get invoked on random elements where tooltip isn't actually shown.